remove: 주석 처리 된 코드 및 사용하지 않는 코드 정리
