+++ src/main/java/Moyoung/Server/auth/userinfo/GoogleUserInfo.java
@@ -13,19 +13,4 @@ public class GoogleUserInfo implements OAuth2UserInfo {
     public String getProviderId() {
         return (String) attributes.get("sub");
     }
-
-    @Override
-    public String getProvider() {
-        return "google";
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getProfileImage() {
-        return (String) attributes.get("picture");
-    }
 }
