+++ src/main/java/Moyoung/Server/auth/userinfo/KakaoUserInfo.java
@@ -13,20 +13,4 @@ public class KakaoUserInfo implements OAuth2UserInfo {
     public String getProviderId() {
         return attributes.get("id").toString();
     }
-
-    @Override
-    public String getProvider() {
-        return "kakao";
-    }
-
-    @Override
-    public String getEmail() {
-        // kakao_account라는 Map에서 추출
-        return (String) ((Map) attributes.get("kakao_account")).get("email");
-    }
-
-    @Override
-    public String getProfileImage() {
-        return (String) ((Map) attributes.get("kakao_account")).get("profile_image");
-    }
 }
