+++ src/main/java/Moyoung/Server/auth/userinfo/NaverUserInfo.java
@@ -13,19 +13,4 @@ public class NaverUserInfo implements OAuth2UserInfo{
     public String getProviderId() {
         return (String) attributes.get("id");
     }
-
-    @Override
-    public String getProvider() {
-        return "naver";
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getProfileImage() {
-        return (String) attributes.get("profile_image");
-    }
 }
