+++ src/main/java/Moyoung/Server/auth/userinfo/OAuth2UserInfo.java
@@ -2,7 +2,4 @@ package Moyoung.Server.auth.userinfo;
 
 public interface OAuth2UserInfo {
     String getProviderId();
-    String getProvider();
-    String getEmail();
-    String getProfileImage();
 }
