+++ src/main/java/Moyoung/Server/crawler/controller/CrawlerController.java
@@ -1,6 +1,6 @@
 package Moyoung.Server.crawler.controller;
 
-import Moyoung.Server.crawler.service.CrawlerServiceV2;
+import Moyoung.Server.crawler.service.CrawlerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +12,7 @@ import java.io.IOException;
 @RestController
 @RequiredArgsConstructor
 public class CrawlerController {
-    private final CrawlerServiceV2 crawlerService;
+    private final CrawlerService crawlerService;
 
     @PostMapping("/manual/re-crawl")
     public ResponseEntity manualReCrawl() throws IOException {
