--- src/main/java/Moyoung/Server/crawler/response/InfoResponse.java
@@ -1,24 +1,0 @@
-//package Moyoung.Server.crawler.response;
-//
-//import com.google.gson.annotations.SerializedName;
-//import lombok.Getter;
-//
-//@Getter
-//public class InfoResponse {
-//    @SerializedName("movieListResult")
-//    private MovieListResult movieListResult;
-//
-//    @Getter
-//    public static class MovieListResult {
-//        @SerializedName("movieList")
-//        private MovieInfo[] movieList;
-//    }
-//
-//    @Getter
-//    public static class MovieInfo {
-//        @SerializedName("genreAlt")
-//        private String genre;
-//        @SerializedName("openDt")
-//        private String releaseDate;
-//    }
-//}
